+++ tests/testthat/test-fetch_lineup.R
@@ -18,6 +18,12 @@ test_that("fetch_lineup_afl works for various inputs", {
   # change comp
   expect_s3_class(fetch_lineup_afl(2020, round_number = 1, comp = "AFLW"), "tbl")
   expect_error(fetch_lineup_afl(2020, round_number = 1, comp = "test"))
+  
+  # Check future year and round
+  current_year <- as.numeric(substr(Sys.Date(),1,4))
+  expect_null(fetch_lineup_afl(current_year + 1, round_number = 1, comp = "AFLM"))
+  expect_null(fetch_line_afl(current_year, round_number = 23, comp = "AFLM"))
+  
 })