+++ NEWS.md
@@ -1,6 +1,7 @@
 # fitzRoy (development version)
 
 * Fixed bug that was preventing Fremantle player details from being returned in `fetch_player_details_footwire` ([#169](https://github.com/jimmyday12/fitzRoy/issues/169))
+* Fixed bug that was causing an error if a round didn't have full lineups
 
 # fitzRoy 1.1.0